Update Arduino_Code.ino
I added the code for homing the SCARA and functions for the storage.
I also added a flag that signals that the storage is empty - if this happens should the robot automatically go retrieve the last node or should we send this signal to the xbee? If we choose the latter, how do we do that? *i just did a serialwrite in line 113.

And removed the below code from retrieve and place because of homing code - let me know if that's wrong.

//Store the new current position
  currentPosition[0] = xend;
  currentPosition[1] = yend;
  currentPosition[2] = zend;

  //End position becomes rest location
  xend = RestPositon[0];
  yend = RestPositon[1];
  zend = RestPositon[2];